+++ src/server.js
@@ -59,7 +59,7 @@ app.use((req, res, next) => {
     'Content-Security-Policy',
     [
       "default-src 'self'",
-      "img-src 'self' data:",
+      "img-src 'self' data: https: http:",
       "style-src 'self' 'unsafe-inline'",
       "script-src 'self' 'unsafe-inline'",
       "connect-src 'self'",